onboard_bolt/requirements.txt
python-dotenv
langchain
langchain-openai
slack_bolt